refactor: 지도 페이지에서 바텀시트 분할
