+++ app/maps/create/index.tsx
@@ -7,8 +7,8 @@ import CaretLeftIcon from '@/assets/icons/caret_left.svg';
 import CaretRightIcon from '@/assets/icons/caret_right.svg';
 import ExitIcon from '@/assets/icons/exit.svg';
 import TelescopeIcon from '@/assets/icons/telescope.svg';
+import CustomBottomSheet from '@/components/common/BottomSheet';
 import { CommonButton } from '@/components/common/Button';
-import CustomBottomSheet from '@/components/features/BottomSheet';
 import { usePost } from '@/contexts/PostContext';
 import { useKeyboardVisible } from '@/hooks/useKeyboardVisible';
 import fetchNearestPlaceName from '@/lib/googlePlaces';
